+++ src/tests/tensorly_torch/inference/general/layers/leaves/parametric/test_multivariate_gaussian.py
@@ -14,7 +14,7 @@ from spflow.tensorly.utils.helper_functions import tl_toNumpy
 tc = unittest.TestCase()
 
 def test_layer_likelihood(do_for_all_backends):
-    torch.set_default_dtype(torch.float64)
+    torch.set_default_dtype(torch.float32)
 
     mean_values = [
         tl.zeros(2),
@@ -51,7 +51,7 @@ def test_layer_likelihood(do_for_all_backends):
     tc.assertTrue(np.allclose(tl_toNumpy(layer_ll), tl_toNumpy(nodes_ll)))
 
 def test_gradient_computation(do_for_all_backends):
-    torch.set_default_dtype(torch.float64)
+    torch.set_default_dtype(torch.float32)
 
     if do_for_all_backends == "numpy":
         return
@@ -177,5 +177,5 @@ def test_update_backend(do_for_all_backends):
 
 
 if __name__ == "__main__":
-    torch.set_default_dtype(torch.float64)
+    torch.set_default_dtype(torch.float32)
     unittest.main()
